Send beat-links logs to our own log file.
This way regardless of where a log message comes from, it can be found
in the proper place, with rollover and size limiting, etc. This is quite
important in trying to help people figure out what is going wrong! And
it only became possible when I upgraded beat-link to use slf4j.