Greatly clean up connection management.
Now, regardless of how a connection is started or stopped (the GUI
checkbox, the connect and disconnect functions, or by Carabiner closing
the connection), it is handled gracefully, and the UI is updated
appropriately (but only if it has been created; it will also create with
current information if a connection is already active when the
configuration window is created).

All that is now missing is a state icon like Triggers have, and the new
Link trigger mode UI and implementation.