Support online toggling.
You can now take BLT online even if it started out offline, and you
can go back offline at any time. This feature is in a new Network
menu, and the Request Metadata option, as well as Ableton Link
Carabiner Connection configuration option have been moved to this
menu.

When enabling metadata, if the device number won't support reliable
queries, a dialog explaining the situation is presented, and offering
the user options for resolution, including bouncing offline and back
on as long as there is a free device number on the network.