Improve expression parsing.
Previously an extra closing paren would cause anything after it to be
silently ignored. Now it will properly complain about an unmatched
delimiter. Parse errors are also now identified with the trigger
number and expression type, as well as proper line and column numbers.
Runtime exceptions also show the trigger number and type description
in the log, with better formatting.