Set up to always work with metadata caches.
In the new model, we always start the MetadataFinder when we are
online, but we leave it in its new passive mode if the user does not
want to actively request metadata for each track loaded. That way we
can at least still track which slots have media in them, and allow
metadata caches to be attached to those slots, thereby making metadata
for that library available even when there are four busy players on
the network.