Get working in Java 9, 10
Add Leiningen flag needed to build, and change to the Java 9 approach
for installing an About handler.

The dark Substance look and feel is incompatible, and crashes, so for
now it is using the standard system L&F, which looks terrible and is
in places illegible. But this is a start; basic functionality seems to
be there.