Get backwards compatibility actually working, #31
This was a lot more complicated than actually expected, because the
code for the Java 9 case as well as the older Mac case both need to be
able to compile on the same machine. Thankfully the magic of `eval` to
defer the compilation to runtime saved the day.

It also seems to be necessary to build the uberjar from Java 8 to
avoid leaking dependencies on Java 9 event classes which cause it to
crash when run in older environments, so commented out the Java 9 line
in project.clj.