Expand Carabiner sync mode concept.
Rather than a Master? toggle, we need a three-state Sync Mode choice,
where the old choices become Off and Triggers, and there is a new Full
choice for bidirectional sync between Ableton Link and Pioneer Pro DJ
Link. This commit gets the first two options working with the new UI;
we need to add an interface to turn on status updates before we can
support letting Link be the tempo master.