Implement Pioneer->Ableton Link ("Passive") sync.
I think the framework is working, but my Link sync in general has gone
to crap, even with the old triggers based approach. I need to get to
the bottom of that. It may be that I am on a wireless network for
development. I will try wired up tomorrow, and check older versions of
BLT that used to work better too.