Send trailing newlines in Carabiner commands to avoid parse errors.
When sending commands too quickly, e.g. when tempo changes in a beat
packet, they are being delivered to Carabiner as a single packet, and
it is causing parse failures. The newline will fix the parse errors,
but a new version of Carabiner will be required to properly process
commands following newline characters within a packet; they are not
detected with the current version.