First crack at Ableton->Pioneer sync implementation.
This needs testing with a wired connection, but I am delighted to
report that in such a setup, the sync in the other direction is
working perfectly! I had simply forgotten how crucial that is when I
was testing last night. You can't get a stable beat alignment on my
wireless network.