Make Load button actually work!
It is currently hardcoded to tell Player 2 to load the track, until I
can add a menu to choose the destination player, but this means
end-to-end track choosing and loading is working! It is just a matter
of implementing the rest of the menu types, adding the destination
menu (with proper handling of players coming and going), and niceties
like good window and element sizing, and warning you when a player is
playing so it cannot load a track.