Add a bunch of help information and issue reporting.
Created a new Help menu, moved the About option to there on
Windows (to be more idiomatic), and added options to open the user
guide, project page, and Gitter chat.

If the current platform supports launching an email client, also added
menu options to compose an issue report (pre-populated with version
details) and a user greeting.