Adapt to the brave, new Crate Digger world!
We can now get metadata in so many more circumstances, we no longer
even ask the user if they want to. We will always try. The choice now
is just whether or not you want to try to use a real player number, to
do things like get metadata for CDs and unanalzyed tracks, and control
the tempo.