Force size of waveform preview component.
Now that it no longer has a fixed size of its own, in order to work
better in the context of the show interface (especially with color
previews that can be up to 1200 pixels wide), we need to force it to
stay the size the Player Status window was designed for.

Also make the Player Status window itself non-resizable since it is
not designed to cope with being any other size than what it chooses,
and now that we fit smaller screens by using a 2x2 arrangement.