WIP attempt at filtering track panels.
Things are getting sorted out of order, and I am anyway realizing that
there is no need to try to walk through and update the components one
by one like this, I can just replace them en masse. But wanted to keep
a record of this code in case there is ever a reason to go back to
something like it and fix it.