WIP: Event management for cue entry/exit, start/stop.
I think the bulk of this is now written, but there are some TODOs to
finish, and *none* of it has been tested at all, so there are
undoubtedly many mistakes to sort out. Once I get some sleep! But the
weekend is coming, and I think I should be able to put out another
build that is ready for people to start trying this stuff then.