Store only cue UUIDs in the interval map.
This has a couple advantages: it lets us repaint a single cue when we
change its color, because the interval map is just a pointer to the
actual cue now, and so repaint always gets the latest values from it.
Previously we had to rebuild the entire track's cue information to
change a single cue color. It also is more efficient when processing
events from the players, because we want to store the set of UUIDs of
the currently-entered cues, and that is now what we get out of the
interval map query with no transformation required.