Don't stutter cues at end.
It turns out that if CDJs send a status packet shortly enough after a
beat packet, they don't always update the beat number in that packet,
even though they have updated the beat-within-bar number. This was
causing BLT to think the cue had been re-entered after the beat packet
exited it, until the next status packet with the correct beat number
arrived.

This change suppresses the effect of a status packet on the show cue
state if that packet arrives within 5 milliseconds after a beat
packet.