Fix opening show filesystems in JDK 13, closes #83
JDK 13 introduced a new method,
`java.nio.file.Filesystems.newFileSystem(Path path,
Map<String, ?> env)`, and sine Beat Link Trigger was passing a `Path`
and `null`, the runtime couldn't distinguish this new method from the
one we had been previously using. Although the old method worked
perfectly well with a `null` `ClassLoader`, this new method crashes
with a `NullPointerException` when given a `null` `Map`.

The fix is to pass a non-`null` class loader, which explicitly chooses
the method that exists in older and newer JDKs, and thus works in
both.