Only build installers on master branch. [skip-ci]
My main goal here is to avoid doing it on the user guide branches,
which are the only ones I regularly create today. It may become unduly
restrictive in the future if I am working on a feature branch, but I
would rather err on the side of being conservative now, and can add
such branches later, or switch to specifically excluding the guide
branches if there end up being lots of branches that deserve
installers.