Split cues window into its own namespace. [skip ci]
Also involved creating a shared show-util namespace for things used by
both it and the show window. The gargantuan shows namespace has needed
this for a while, and that will help me as I undertake the big new
effort to implement linked cues, which will also be used by the new
phrase triggers.