More progress on supporting phrase trigger cues, #119 [skip ci]
Fixed a whole mess of bugs I had introduced which became evident when
I started trying to manipulate cues in phrase triggers (and even in
tracks, especially at the end of the track). The worst are behind us,
I think,  but things are still a little too glitchy for me to build a
preview to share with anyone yet. Hopefully after dinner I will finish
off a few more items on my checklist, and feel ready for that!