Relate section size boundaries and cue boundaries, #119 [skip ci]
You can now no longer shrink a section past existing cues, and the
spinner UI updates immediately to reflect that, and the cue end
spinners update immediately when the section size changes in a way
that affects them too.

Also, the sections are repainted in both windows when they change.

Sadly, there is still a bit more flakiness I want to solve before
building a new preview, but I need to get to bed!