Resolve phrase UUID conflicts when opening shows, #119
If users create show files by starting with copies of other show
files, they might well end up having phrases with the same UUIDs,
which violates an important invariant in the show code. We can solve
this by reassigning the UUIDs of any conflicting phrases in the show
being opened, as long as we maintain internal consistency within that
show, which will be easier to do before building the GUI and runtime
structures for it.