Fix display of player position in phrase triggers, #119
Due to floating-point issues it was jumping spastically back and forth
between beats. Also, once I got it looking right for the cues editor,
I realized it would be easier to just have it look the same way in the
phrase trigger preview row in the show window, than explaining the
completely new lower-resolution beat-level position display in the
user guide. So I did, and that is a much better user experience as it
turns out.