Remove unused dependency on humanize
There was a release version that was newer, but this one was also
pulling in an old version of clojure.math/numeric-tower, which
produced a warning about `abs` being shadowed. Getting rid of it
showed we were relying on that implicit dependency in some code, but
we can replace that with the new, faster, tinier clojure.math in the
core library.