Stop trying to build hosted opus user guide
It can't build from the development branch and still have a version
number that works with the embedded guide. If we want to start having
hosted guides for that branch before merging it to main, we will need
to create a guide-opus tracking branch the way we do for other
releases.