Start converting to nicer Java interop offered by Clojure 1.12
I'll pause on this until clj-kondo can catch up, it choked immediately
on the new array syntax, but it looks like that has been implemented
and will be released in not too long.

For other things that can be improved, a good field guide is:
https://www.jarrodctaylor.com/posts/Clojure-1.12-Field-Guide/

As part of this project, let's also clean up any reflection warnings
that `lein check` can find.